requirements.txt

numpy
pandas
scikit-learn
sentence-transformers
neat-python
langchain-ollama
tqdm
pyyaml
openpyxl
